Rename files (phase I).
