+++ src/tree/containers/files/DirectoryTreeItem.ts
@@ -1,7 +1,7 @@
 import * as vscode from 'vscode';
 import { AzExtParentTreeItem, AzExtTreeItem, IActionContext } from "vscode-azureextensionui";
 import { DirectoryItem } from "../../../docker/DockerContainerDirectoryProvider";
-import { DockerUri } from '../../../docker/files/dockerUri';
+import { DockerUri } from '../../../docker/files/DockerUri2';
 import { AzExtParentTreeItemIntermediate } from '../../AzExtParentTreeItemIntermediate';
 import { FileTreeItem } from "./FileTreeItem";
